Setting up instrumentation hookups
