+++ lib/ripple/instrumentation/backend/excon.rb
@@ -1,0 +1,8 @@
+require 'ripple/instrumentation/http'
+require 'riak/client/excon_backend'
+
+module Riak::Client
+  class ExconBackend
+    include Ripple::Instrumentation::HTTP
+  end
+end
