+++ lib/ripple/instrumentation/backend/net_http.rb
@@ -1,0 +1,8 @@
+require 'ripple/instrumentation/http'
+require 'riak/client/net_http_backend'
+
+module Riak::Client
+  class NetHTTPBackend
+    include Ripple::Instrumentation::HTTP
+  end
+end
